Add NoEvents, Loading, Statistics. Code review
