+++ src/components/edit-event.js
@@ -1,0 +1,8 @@
+const createEditEventTemplate = () => {
+  return (
+    `<form class="event  event--edit" action="#" method="post">
+    </form>`
+  );
+};
+
+export {createEditEventTemplate};
