--- src/components/edit-event/edit-event-header/edit-event-time.js
@@ -1,28 +1,0 @@
-const createEditEventTimeTemplate = () => {
-  return (
-    `<div class="event__field-group  event__field-group--time">
-      <label class="visually-hidden" for="event-start-time-1">
-        From
-      </label>
-      <input
-        class="event__input  event__input--time"
-        id="event-start-time-1"
-        type="text"
-        name="event-start-time"
-        value="18/03/19 00:00"
-      >
-      &mdash;
-      <label class="visually-hidden" for="event-end-time-1">
-        To
-      </label>
-      <input
-        class="event__input  event__input--time"
-        id="event-end-time-1"
-        type="text" name="event-end-time"
-        value="18/03/19 00:00"
-      >
-    </div>`
-  );
-};
-
-export {createEditEventTimeTemplate};
