+++ src/components/loading.js
@@ -1,0 +1,9 @@
+const createLoadingTemplate = () => {
+  return (
+    `<p class="trip-events__msg">
+      Loading...
+    </p>`
+  );
+};
+
+export {createLoadingTemplate};
