+++ src/components/no-events.js
@@ -1,0 +1,9 @@
+const createNoEventsTemplate = () => {
+  return (
+    `<p class="trip-events__msg">
+      Click New Event to create your first point
+    </p>`
+  );
+};
+
+export {createNoEventsTemplate};
