+++ src/components/statistics.js
@@ -1,0 +1,23 @@
+const createStatisticsTemplate = () => {
+  return (
+    `<section class="statistics">
+      <h2 class="visually-hidden">
+        Trip statistics
+      </h2>
+      <div class="statistics__item statistics__item--money">
+        <canvas class="statistics__chart  statistics__chart--money" width="900">
+        </canvas>
+      </div>
+      <div class="statistics__item statistics__item--transport">
+        <canvas class="statistics__chart  statistics__chart--transport" width="900">
+        </canvas>
+      </div>
+      <div class="statistics__item statistics__item--time-spend">
+        <canvas class="statistics__chart  statistics__chart--time" width="900">
+        </canvas>
+      </div>
+    </section>`
+  );
+};
+
+export {createStatisticsTemplate};
