+++ src/components/trip-tabs.js
@@ -1,8 +1,12 @@
 const createTripTabsTemplate = () => {
   return (
     `<nav class="trip-controls__trip-tabs  trip-tabs">
-      <a class="trip-tabs__btn  trip-tabs__btn--active" href="#">Table</a>
-      <a class="trip-tabs__btn" href="#">Stats</a>
+      <a class="trip-tabs__btn  trip-tabs__btn--active" href="#">
+        Table
+      </a>
+      <a class="trip-tabs__btn" href="#">
+        Stats
+      </a>
     </nav>`
   );
 };
